fix(ios): satisfy keyboard handoff lint limits
